Changes to be committed: 	modified:   10-Enhanced-Reporting-Functions.ps1
HTML Reports:
Change title date to show 'As of' that reflects the date of the data file.
Change the footer to show the date the report was generated plus the dat the data was generated.

Excel Report:
Added date that shows the 'As of' date the data was generated.

For both report the date is passed on the parameter -DataDate.
If not supplied it defaults to the current date/time.
If passing in the results from a stored data file this could be inaccurate.
You should retriev ethe LastWriteTime fromthe file and pass that into the report functions.